Rename service dir to services
